% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticache_operations.R
\name{create_snapshot}
\alias{create_snapshot}
\title{Creates a copy of an entire cluster or replication group at a specific moment in time}
\usage{
create_snapshot(
  ReplicationGroupId = "string",
  CacheClusterId = "string",
  SnapshotName = "string"
)
}
\arguments{
\item{ReplicationGroupId}{The identifier of an existing replication group. The snapshot is created from this replication group.}

\item{CacheClusterId}{The identifier of an existing cluster. The snapshot is created from this cluster.}

\item{SnapshotName}{[required] A name for the snapshot being created.}
}
\description{
Creates a copy of an entire cluster or replication group at a specific moment in time.
}
\details{
This operation is valid for Redis only.
}
\examples{

}
