% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticache_operations.R
\name{delete_cache_parameter_group}
\alias{delete_cache_parameter_group}
\title{Deletes the specified cache parameter group}
\usage{
delete_cache_parameter_group(CacheParameterGroupName)
}
\arguments{
\item{CacheParameterGroupName}{[required] The name of the cache parameter group to delete.

The specified cache security group must not be associated with any clusters.}
}
\description{
Deletes the specified cache parameter group. You cannot delete a cache parameter group if it is associated with any cache clusters.
}
\section{Accepted Parameters}{
\preformatted{delete_cache_parameter_group(
  CacheParameterGroupName = "string"
)
}
}

