% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticache_operations.R
\name{delete_snapshot}
\alias{delete_snapshot}
\title{Deletes an existing snapshot}
\usage{
delete_snapshot(SnapshotName)
}
\arguments{
\item{SnapshotName}{The name of the snapshot to be deleted.}
}
\description{
Deletes an existing snapshot. When you receive a successful response from this operation, ElastiCache immediately begins deleting the snapshot; you cannot cancel or revert this operation.
}
\details{
This operation is valid for Redis only.
}
\examples{

}
