% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticache_operations.R
\name{describe_cache_subnet_groups}
\alias{describe_cache_subnet_groups}
\title{Returns a list of cache subnet group descriptions}
\usage{
describe_cache_subnet_groups(
  CacheSubnetGroupName = "string",
  MaxRecords = 123,
  Marker = "string"
)
}
\arguments{
\item{CacheSubnetGroupName}{The name of the cache subnet group to return details for.}

\item{MaxRecords}{The maximum number of records to include in the response. If more records exist than the specified \code{MaxRecords} value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.}

\item{Marker}{An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by \code{MaxRecords}.}
}
\description{
Returns a list of cache subnet group descriptions. If a subnet group name is specified, the list contains only the description of that group.
}
\examples{

}
