% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticache_operations.R
\name{modify_cache_parameter_group}
\alias{modify_cache_parameter_group}
\title{Modifies the parameters of a cache parameter group}
\usage{
modify_cache_parameter_group(CacheParameterGroupName, ParameterNameValues)
}
\arguments{
\item{CacheParameterGroupName}{The name of the cache parameter group to modify.}

\item{ParameterNameValues}{An array of parameter names and values for the parameter update. You must supply at least one parameter name and value; subsequent arguments are optional. A maximum of 20 parameters may be modified per request.}
}
\description{
Modifies the parameters of a cache parameter group. You can modify up to 20 parameters in a single request by submitting a list parameter name and value pairs.
}
\examples{

}
