% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticache_operations.R
\name{modify_cache_subnet_group}
\alias{modify_cache_subnet_group}
\title{Modifies an existing cache subnet group}
\usage{
modify_cache_subnet_group(
  CacheSubnetGroupName = "string",
  CacheSubnetGroupDescription = "string",
  SubnetIds = list(
    "string"
  )
)
}
\arguments{
\item{CacheSubnetGroupName}{[required] The name for the cache subnet group. This value is stored as a lowercase string.

Constraints: Must contain no more than 255 alphanumeric characters or hyphens.

Example: \code{mysubnetgroup}}

\item{CacheSubnetGroupDescription}{A description of the cache subnet group.}

\item{SubnetIds}{The EC2 subnet IDs for the cache subnet group.}
}
\description{
Modifies an existing cache subnet group.
}
\examples{

}
