% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticache_operations.R
\name{modify_replication_group_shard_configuration}
\alias{modify_replication_group_shard_configuration}
\title{Modifies a replication group's shards (node groups) by allowing you to add shards, remove shards, or rebalance the keyspaces among exisiting shards}
\usage{
modify_replication_group_shard_configuration(
  ReplicationGroupId = "string",
  NodeGroupCount = 123,
  ApplyImmediately = TRUE|FALSE,
  ReshardingConfiguration = list(
    list(
      NodeGroupId = "string",
      PreferredAvailabilityZones = list(
        "string"
      )
    )
  ),
  NodeGroupsToRemove = list(
    "string"
  ),
  NodeGroupsToRetain = list(
    "string"
  )
)
}
\arguments{
\item{ReplicationGroupId}{[required] The name of the Redis (cluster mode enabled) cluster (replication group) on which the shards are to be configured.}

\item{NodeGroupCount}{[required] The number of node groups (shards) that results from the modification of the shard configuration.}

\item{ApplyImmediately}{[required] Indicates that the shard reconfiguration process begins immediately. At present, the only permitted value for this parameter is \code{true}.

Value: true}

\item{ReshardingConfiguration}{Specifies the preferred availability zones for each node group in the cluster. If the value of \code{NodeGroupCount} is greater than the current number of node groups (shards), you can use this parameter to specify the preferred availability zones of the cluster's shards. If you omit this parameter ElastiCache selects availability zones for you.

You can specify this parameter only if the value of \code{NodeGroupCount} is greater than the current number of node groups (shards).}

\item{NodeGroupsToRemove}{If the value of \code{NodeGroupCount} is less than the current number of node groups (shards), the \code{NodeGroupsToRemove} or \code{NodeGroupsToRetain} is a required list of node group ids to remove from or retain in the cluster.

ElastiCache for Redis will attempt to remove all node groups listed by \code{NodeGroupsToRemove} from the cluster.}

\item{NodeGroupsToRetain}{If the value of \code{NodeGroupCount} is less than the current number of node groups (shards), the \code{NodeGroupsToRemove} or \code{NodeGroupsToRetain} is a required list of node group ids to remove from or retain in the cluster.

ElastiCache for Redis will attempt to remove all node groups except those listed by \code{NodeGroupsToRetain} from the cluster.}
}
\description{
Modifies a replication group's shards (node groups) by allowing you to add shards, remove shards, or rebalance the keyspaces among exisiting shards.
}
\examples{

}
