% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticache_package.R
\docType{package}
\name{paws.elasticache-package}
\alias{paws.elasticache}
\alias{paws.elasticache-package}
\title{paws.elasticache: Amazon ElastiCache}
\description{
Amazon ElastiCache is a web service that makes it easier to
set up, operate, and scale a distributed cache in the cloud.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
