% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticache_operations.R
\name{revoke_cache_security_group_ingress}
\alias{revoke_cache_security_group_ingress}
\title{Revokes ingress from a cache security group}
\usage{
revoke_cache_security_group_ingress(CacheSecurityGroupName,
  EC2SecurityGroupName, EC2SecurityGroupOwnerId)
}
\arguments{
\item{CacheSecurityGroupName}{The name of the cache security group to revoke ingress from.}

\item{EC2SecurityGroupName}{The name of the Amazon EC2 security group to revoke access from.}

\item{EC2SecurityGroupOwnerId}{The AWS account number of the Amazon EC2 security group owner. Note that this is not the same thing as an AWS access key ID - you must provide a valid AWS account number for this parameter.}
}
\description{
Revokes ingress from a cache security group. Use this operation to disallow access from an Amazon EC2 security group that had been previously authorized.
}
\examples{

}
