% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticache_operations.R
\name{test_failover}
\alias{test_failover}
\title{Represents the input of a \code{TestFailover} operation which test automatic failover on a specified node group (called shard in the console) in a replication group (called cluster in the console)}
\usage{
test_failover(ReplicationGroupId, NodeGroupId)
}
\arguments{
\item{ReplicationGroupId}{The name of the replication group (console: cluster) whose automatic failover is being tested by this operation.}

\item{NodeGroupId}{The name of the node group (called shard in the console) in this replication group on which automatic failover is to be tested. You may test automatic failover on up to 5 node groups in any rolling 24-hour period.}
}
\description{
Represents the input of a \code{TestFailover} operation which test automatic failover on a specified node group (called shard in the console) in a replication group (called cluster in the console).
}
\details{
\strong{Note the following}
\itemize{
\item A customer can use this operation to test automatic failover on up to 5 shards (called node groups in the ElastiCache API and AWS CLI) in any rolling 24-hour period.
\item If calling this operation on shards in different clusters (called replication groups in the API and CLI), the calls can be made concurrently.
\item If calling this operation multiple times on different shards in the same Redis (cluster mode enabled) replication group, the first node replacement must complete before a subsequent call can be made.
\item To determine whether the node replacement is complete you can check Events using the Amazon ElastiCache console, the AWS CLI, or the ElastiCache API. Look for the following automatic failover related events, listed here in order of occurrance:
}
\enumerate{
\item Replication group message: \code{Test Failover API called for node group <node-group-id>}
\item Cache cluster message: \code{Failover from master node <primary-node-id> to replica node <node-id> completed}
\item Replication group message: \code{Failover from master node <primary-node-id> to replica node <node-id> completed}
\item Cache cluster message: \code{Recovering cache nodes <node-id>}
\item Cache cluster message: \code{Finished recovery for cache nodes <node-id>}
}

For more information see:
\itemize{
\item \href{http://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/ECEvents.Viewing.html}{Viewing ElastiCache Events} in the \emph{ElastiCache User Guide}
\item \href{http://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_DescribeEvents.html}{DescribeEvents} in the ElastiCache API Reference
}

Also see, \href{http://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/AutoFailover.html#auto-failover-test}{Testing Multi-AZ with Automatic Failover} in the \emph{ElastiCache User Guide}.
}
\examples{

}
